создание опроса 8
выделила js, отвечающий за модалку выбора типа вопроса в отдельный модуль, а то в коде стало невозможно ориентироваться
