+++ src/app/common_static/css/create_poll_page.css
@@ -21,7 +21,7 @@ header h1 {
 
 }
 
-.form> *, .chooseQuestionType2 {
+.form> *, .chooseQuestionType {
     display: block;
     width: 100%;
     margin: auto;
