+++ src/app/create_poll_page/templates/create_poll_page.html
@@ -81,7 +81,7 @@
 
 
             </div>
-            <script src="{% static 'scripts/create_poll_page.js' %}"></script>
+            <script src="{% static 'scripts/create_poll_page.js' %}" type="module"></script>
 
         </main>
